return False

+++ 3-deploy_web_static.py
@@ -20,12 +20,12 @@ def do_pack():
     try:
         local(f"tar -cvzf {file_name} web_static")
         return file_name
-    except:
+    except Exception:
         return None
 
 
 def do_deploy(archive_path):
-    """ deploy on server web-01 and web-02"""
+    """deploy on server web-01 and web-02"""
 
     if exists(archive_path) is False:
         return False
@@ -33,16 +33,33 @@ def do_deploy(archive_path):
         achive_file = archive_path.split("/")[-1]
         achive_file_name = achive_file.split(".")[0]
         path = "/data/web_static/releases/"
-        put(local_path=archive_path, remote_path="/tmp/")
-        run(f"mkdir -p /data/web_static/releases/{achive_file_name}")
-        run(f"tar -xzf /tmp/{achive_file} -C {path}{achive_file_name}/")
-        run(f"mv {path}{achive_file_name}/web_static/* \
+        result = put(local_path=archive_path, remote_path="/tmp/")
+        if result.failed:
+            return False
+        result = run(f"mkdir -p /data/web_static/releases/{achive_file_name}")
+        if result.failed:
+            return False
+        result = run(f"tar -xzf /tmp/{achive_file} -C \
             {path}{achive_file_name}/")
-        run(f"rm -rf {path}{achive_file_name}/web_static")
-        run(f"rm -rf /data/web_static/current")
-        run(f"ln -s {path}{achive_file_name}/ /data/web_static/current")
+        if result.failed:
+            return False
+        result = run(f"mv {path}{achive_file_name}/web_static/* \
+            {path}{achive_file_name}/")
+        if result.failed:
+            return False
+        result = run(f"rm -rf {path}{achive_file_name}/web_static")
+        if result.failed:
+            return False
+        result = run("rm -rf /data/web_static/current")
+        if result.failed:
+            return False
+        result = run(f"ln -s {path}{achive_file_name}/\
+            /data/web_static/current")
+        if result.failed:
+            return False
         print("New version deployed!")
-    except:
+        return True
+    except Exception:
         return False
